api/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Redis
redis[hiredis]==5.2.1

# HTTP Client
httpx==0.28.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Document Processing
openpyxl==3.1.5
pdfplumber==0.11.4
python-docx==1.1.2
pytesseract==0.3.13
Pillow==11.1.0

# Validation & Settings
pydantic==2.10.4
pydantic-settings==2.7.1

# Logging
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx  # also used for TestClient

# Linting
ruff==0.8.6
